+++ src/main/java/courgette/runtime/event/EventSubscriber.java
@@ -1,0 +1,22 @@
+package courgette.runtime.event;
+
+import java.beans.PropertyChangeEvent;
+
+public abstract class EventSubscriber {
+
+    protected EventHolder eventHolder;
+
+    protected boolean matchesEvent(CourgetteEvent courgetteEvent) {
+        return eventHolder.getCourgetteEvent().equals(courgetteEvent);
+    }
+
+    protected abstract void sendEvent(EventSender eventSender);
+
+    protected void propertyChange(PropertyChangeEvent evt) {
+        if (evt.getPropertyName().equals("eventHolder")) {
+            this.eventHolder = (EventHolder) evt.getNewValue();
+        } else {
+            sendEvent((EventSender) evt.getNewValue());
+        }
+    }
+}