+++ src/main/java/courgette/runtime/event/subscriber/TestPassedAfterRerunEventSubscriber.java
@@ -1,0 +1,23 @@
+package courgette.runtime.event.subscriber;
+
+import courgette.runtime.event.CourgetteEvent;
+import courgette.runtime.event.EventSender;
+import courgette.runtime.event.EventSubscriber;
+
+import java.beans.PropertyChangeEvent;
+import java.beans.PropertyChangeListener;
+
+public class TestPassedAfterRerunEventSubscriber extends EventSubscriber implements PropertyChangeListener {
+
+    @Override
+    public void sendEvent(EventSender eventSender) {
+        if (matchesEvent(CourgetteEvent.TEST_PASSED_AFTER_RERUN)) {
+            eventSender.send(eventHolder);
+        }
+    }
+
+    @Override
+    public void propertyChange(PropertyChangeEvent evt) {
+        super.propertyChange(evt);
+    }
+}